// src/shared/generated/BrowserFontRuntimePayloadRecord.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { BrowserFontRuntimeFaceRecord } from "./BrowserFontRuntimeFaceRecord";
import type { BrowserFontSettingsRecord } from "./BrowserFontSettingsRecord";

export type BrowserFontRuntimePayloadRecord = { settings: BrowserFontSettingsRecord, faces: Array<BrowserFontRuntimeFaceRecord>, };